requirements.txt
Flask==3.0.0
Pillow==10.1.0
pytesseract==0.3.10
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.0
google-generativeai==0.3.1
protobuf==4.25.1
Werkzeug==3.0.1
requests==2.31.0
numpy==1.26.2
opencv-python==4.8.1.78
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
grpcio==1.60.0
grpcio-status==1.60.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
rsa==4.9
six==1.16.0
urllib3==2.1.0